+++ ide_image_scripts/test/integration/ssh/30-copy-ssh-configs.sh
@@ -26,9 +26,4 @@ Host git.ai-traders.com
 User git
 Port 2222
 IdentityFile ${ide_home}/.ssh/id_rsa
-
-Host gitlab.ai-traders.com
-User git
-Port 2222
-IdentityFile ${ide_home}/.ssh/id_rsa
 " > "${ide_home}/.ssh/config"